refactor to create GetAttachments module
